requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.15.0
yfinance>=0.2.18
scikit-learn>=1.1.0
xgboost>=1.7.0
ta>=0.10.2
google-generativeai>=0.3.0
pydantic>=2.0.0
joblib>=1.3.0
